requirements.txt
piexif
pillow
pillow-heif
pytest
